fix: conflict between omniauth_registrations_controller_extends and OmniauthRegistrationsControllerOverride
